separate range checks witness
