+++ stwo_cairo_prover/crates/prover/src/cairo_air/blake/air.rs
@@ -11,13 +11,13 @@ use stwo_prover::core::pcs::TreeVec;
 use tracing::{span, Level};
 
 use crate::cairo_air::air::CairoInteractionElements;
-use crate::cairo_air::range_checks_air::RangeChecksClaimGenerator;
 use crate::components::utils::TreeBuilder;
 use crate::components::{
     blake_g, blake_round, blake_round_sigma, memory_address_to_id, memory_id_to_big, triple_xor_32,
     verify_bitwise_xor_12, verify_bitwise_xor_4, verify_bitwise_xor_7, verify_bitwise_xor_8,
     verify_bitwise_xor_9,
 };
+use crate::witness::range_checks::RangeChecksClaimGenerator;
 
 #[derive(Serialize, Deserialize, CairoSerialize)]
 pub struct BlakeContextClaim {
