+++ stwo_cairo_prover/crates/prover/src/witness/mod.rs
@@ -1,3 +1,4 @@
 pub mod builtins;
 pub mod cairo;
 pub mod opcodes;
+pub mod range_checks;
